修改内容： 1.回款跟踪 2.生产进度 3.报表 4.定位格式更新 5.加载通知

+++ app/src/main/java/com/cimcitech/cimcly/bean/opport_unit/OpportUnitVo.java
@@ -1,5 +1,6 @@
 package com.cimcitech.cimcly.bean.opport_unit;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -364,7 +365,7 @@ public class OpportUnitVo {
                 private Object ordersource;
                 private Object orgid;
                 private String paymentmethod;
-                private int planmoney;
+                private BigDecimal planmoney;
                 private long plansigndate;
                 private String possibility;
                 private String possibilityValue;
@@ -548,11 +549,11 @@ public class OpportUnitVo {
                     this.paymentmethod = paymentmethod;
                 }
 
-                public int getPlanmoney() {
+                public BigDecimal getPlanmoney() {
                     return planmoney;
                 }
 
-                public void setPlanmoney(int planmoney) {
+                public void setPlanmoney(BigDecimal planmoney) {
                     this.planmoney = planmoney;
                 }
